TST/REF: collect tests by method
