+++ pandas/tests/series/indexing/test_setitem.py
@@ -1,7 +1,9 @@
+from datetime import date
+
 import numpy as np
 import pytest
 
-from pandas import MultiIndex, NaT, Series, date_range, period_range
+from pandas import MultiIndex, NaT, Series, Timestamp, date_range, period_range
 import pandas.testing as tm
 
 
@@ -27,6 +29,26 @@ class TestSetitemDT64Values:
         expected = result.copy()
         result.loc[[]] = 0
         tm.assert_series_equal(result, expected)
+
+    def test_setitem_with_string_index(self):
+        # GH#23451
+        ser = Series([1, 2, 3], index=["Date", "b", "other"])
+        ser["Date"] = date.today()
+        assert ser.Date == date.today()
+        assert ser["Date"] == date.today()
+
+    def test_setitem_with_different_tz_casts_to_object(self):
+        # GH#24024
+        ser = Series(date_range("2000", periods=2, tz="US/Central"))
+        ser[0] = Timestamp("2000", tz="US/Eastern")
+        expected = Series(
+            [
+                Timestamp("2000-01-01 00:00:00-05:00", tz="US/Eastern"),
+                Timestamp("2000-01-02 00:00:00-06:00", tz="US/Central"),
+            ],
+            dtype=object,
+        )
+        tm.assert_series_equal(ser, expected)
 
 
 class TestSetitemPeriodDtype:
